Extract searchResponseAdaptor for reuse
