+++ src/common/dataAdaptors/index.js
@@ -5,6 +5,7 @@ export * from './introAdaptors'
 export * from './mediaAdaptors'
 export * from './objectsToIdsAdaptor'
 export * from './pageAdaptors'
+export * from './searchResponseAdaptors'
 export * from './siteAdaptors'
 export * from './songAdaptors'
 export * from './storyAdaptors'
