+++ src/common/dataHooks/useSearchLoader.js
@@ -6,22 +6,8 @@ import PropTypes from 'prop-types'
 // FPCC
 import useIntersectionObserver from 'common/hooks/useIntersectionObserver'
 import api from 'services/api'
-import {
-  SEARCH,
-  TYPE_PHRASE,
-  TYPE_SONG,
-  TYPE_STORY,
-  TYPE_WORD,
-  TYPE_AUDIO,
-  TYPE_IMAGE,
-  TYPE_VIDEO,
-} from 'common/constants'
-import {
-  basicDatesAdaptor,
-  storySummaryAdaptor,
-  songSummaryAdaptor,
-  mediaAdaptor,
-} from 'common/dataAdaptors'
+import { SEARCH } from 'common/constants'
+import { searchResponseAdaptor } from 'common/dataAdaptors'
 
 /**
  * Calls search API and provides search results and infinite scroll info.
@@ -29,65 +15,6 @@ import {
 function useSearchLoader({ searchParams }) {
   const { sitename } = useParams()
   const searchParamString = searchParams.toString()
-
-  const pagesDataAdaptor = (pages) =>
-    pages.map((page, index) => singlePageDataAdaptor(page, index))
-
-  const singlePageDataAdaptor = (page, index) => {
-    const formattedEntries = page?.results?.map((result) =>
-      resultAdaptor(result),
-    )
-    return {
-      ...page,
-      pageNumber: index + 1,
-      results: formattedEntries,
-    }
-  }
-
-  const resultAdaptor = (result) => {
-    const baseObject = {
-      id: result?.entry?.id,
-      title: result?.entry?.title,
-      type: result?.type,
-      sitename: result?.entry?.site?.slug,
-      siteTitle: result?.entry?.site?.title,
-      visibility: result?.entry?.visibility,
-      siteVisibility: result?.entry?.site?.visibility,
-      ...basicDatesAdaptor({ item: result?.entry }),
-    }
-    switch (result?.type) {
-      case TYPE_WORD:
-      case TYPE_PHRASE:
-        return {
-          ...baseObject,
-          translations: result?.entry?.translations || [],
-          audio: result?.entry?.relatedAudio || [],
-          image: result?.entry?.relatedImages?.[0] || null,
-        }
-      case TYPE_SONG:
-        return {
-          ...songSummaryAdaptor({ item: result?.entry }),
-          ...baseObject,
-        }
-      case TYPE_STORY:
-        return {
-          ...storySummaryAdaptor({ item: result?.entry }),
-          ...baseObject,
-        }
-      case TYPE_AUDIO:
-      case TYPE_IMAGE:
-      case TYPE_VIDEO:
-        return {
-          ...mediaAdaptor({ type: result?.type, data: result?.entry }),
-          ...baseObject,
-        }
-      default:
-        return {
-          ...baseObject,
-          message: 'Unrecognized entry type!',
-        }
-    }
-  }
 
   const queryKeySite = sitename || 'cross-site'
   const queryFn = sitename
@@ -107,7 +34,7 @@ function useSearchLoader({ searchParams }) {
     refetchOnWindowFocus: false,
     refetchOnReconnect: false,
     select: (responseData) => ({
-      pages: pagesDataAdaptor(responseData.pages),
+      pages: searchResponseAdaptor(responseData),
       pageParams: responseData.pageParams,
     }),
   })
